+++ source/extensions/isaacsim.core.experimental.objects/python/impl/shapes/cylinder.py
@@ -50,7 +50,9 @@ class Cylinder(Shape):
             If the input shape is smaller than expected, data will be broadcasted (following NumPy broadcast rules).
         axes: Axes (cylinder's axis along which the spine is aligned) (shape ``(N,)``).
             If the input shape is smaller than expected, data will be broadcasted (following NumPy broadcast rules).
-        colors: Display colors (shape ``(N, 3)``).
+        colors: Normalized RGB display colors (shape ``(N, 3)``) or case-insensitive string representations.
+            Supported string representations include hex codes and X11/CSS4 color names without spaces,
+            as well as any other format supported by Matplotlib. Alpha channel is ignored for string representations.
             If the input shape is smaller than expected, data will be broadcasted (following NumPy broadcast rules).
         positions: Positions in the world frame (shape ``(N, 3)``).
             If the input shape is smaller than expected, data will be broadcasted (following NumPy broadcast rules).
@@ -65,6 +67,7 @@ class Cylinder(Shape):
 
     Raises:
         ValueError: If resulting paths are mixed (existing and non-existing prims) or invalid.
+        ValueError: Invalid string representation format for the colors.
         AssertionError: If wrapped prims are not USD Cylinder.
         AssertionError: If both positions and translations are specified.
 
@@ -77,7 +80,7 @@ class Cylinder(Shape):
         >>> # given an empty USD stage with the /World Xform prim,
         >>> # create yellow cylinders at paths: /World/prim_0, /World/prim_1, and /World/prim_2
         >>> paths = ["/World/prim_0", "/World/prim_1", "/World/prim_2"]
-        >>> prims = Cylinder(paths, colors=[1.0, 1.0, 0.0])  # doctest: +NO_CHECK
+        >>> prims = Cylinder(paths, colors="#ff0")  # doctest: +NO_CHECK
     """
 
     def __init__(
@@ -89,7 +92,7 @@ class Cylinder(Shape):
         heights: float | list | np.ndarray | wp.array | None = None,
         axes: Literal["X", "Y", "Z"] | list[Literal["X", "Y", "Z"]] | None = None,
         # Shape
-        colors: list | np.ndarray | wp.array | None = None,
+        colors: str | list | np.ndarray | wp.array | None = None,
         # XformPrim
         positions: list | np.ndarray | wp.array | None = None,
         translations: list | np.ndarray | wp.array | None = None,