+++ source/extensions/isaacsim.core.experimental.objects/python/tests/test_mesh.py
@@ -304,3 +304,22 @@ class TestMesh(omni.kit.test.AsyncTestCase):
                 check_lists(expected_v0, output[0])
                 check_lists(expected_v1, output[1])
                 check_lists(expected_v2, output[2])
+
+    @parametrize(backends=["usd"], prim_class=Mesh, populate_stage_func=populate_stage)
+    async def test_display_colors(self, prim, num_prims, device, backend):
+        choices = [
+            (0.1, 0.2, 0.3),  # RGB tuple
+            "#aBc",  # case-insensitive short hex RGB
+            "#0A1b2C",  # case-insensitive hex RGB
+            "0.5",  # grayscale
+            "k",  # basic color
+            "AquaMarine",  # case-insensitive X11/CSS4 color with no spaces
+            "xkcd:eggShell",  # case-insensitive  xkcd color
+            "tab:Green",  # case-insensitive tableau color
+            "C2",  # CN color specification
+            "none",  # special value (fully transparent)
+        ]
+        for indices, expected_count in draw_indices(count=num_prims, step=2):
+            cprint(f"  |    |-- indices: {type(indices).__name__}, expected_count: {expected_count}")
+            for v0, expected_v0 in draw_choice(shape=(expected_count,), choices=choices):
+                prim.set_display_colors(v0, indices=indices)